[ADD] base: fetch translations online
Fetch the latest translations from an external provider (e.g. nightly/CDN
server) instead of relying on local files. This would allow to:
- fetch translations more regularly than once per week
- be able to update translations without having to wait for the source code to
  be updated
- reduce the size of the repository/install files by removing .po files from
  this repository

Changes in the way the translations are loaded.
Instead of having public methods that does several unrelated things, split in
small private methods with only one purpose:
* load_lang: no longer activates or create a language, one method
  _activate_lang, one method _create_lang

* trans_load_data: no longer activates the language if inactive, should be
  tested and activated explicitly

* import_lang: used to create a new language if it was not present or
  activated. Will now raise an error.

[FIX] reintroduce module_name parameter

This reverts commit 62d3675869048
Actually maybe we need it to be able to find pot associated to a po
file
Don't use it for code translation though (that one is still true)

Create a translation reader for the pot file if exists
This is a small violation of the reader api as not working for csv files