[FIX] stock: change test cases accordingly UOM precision change
Purpose of this commit is in master decimal value of qty is
coverted base on decimal accuracy for UOM(i.e digits=3)
i.e specified on field level so it's doen't convert it base
on uom rounding, Now in our case we removed decimal accuracy
for UOM and we are changing the decimal value of qty based on
uom rounding i.e specified on field level, it will make the
result different, So we need to change the test case.

For EX In master
	Product unit- rounding 0.001 (6 unit on hand)
	digit = 3
	dozen rounding = 1
So when we input 0.5 with dozen it will stay as 0.5 as we
have not yet converted.In our case it will converted 0.5 to 0
qty because of we coverted the decimal value at field level
based on uom_field rounding so need to change the uom rounding
of donzen to 0.001 and base on that change the next steps of
test case.

Removed some code from stock move line because currently
in master precision_digits is decimal accuracy for
UOM(i.e digits=3) and precision_rounding is product uom rounding
so both are different but in our case we removed decimal accuracy
for UOM so rounding is done with same uom and result will be always
same so no need to check that case.

task-1914025