[IMP] website_forum: show share buttons only on validated question
Before this commit:
when a user asks a question on a forum, share buttons are always showing even if
question is validated or not and hence it will lead to 404 error.

After this commit:
when a user asks a question on a forum, share buttons will only show after a
question is validated.

task-2064874