[FIX] service: use collections.deque for queue
In the PreforkServer, the python started process is responsible to
forward all incoming system signals to the forked processes. Signals
received are registered in a queue and later consumed.

The queue was implemented on the python list which implements a stack,
removing the first element from a list is a O(n) operation. The list has
been replaced by a `collections.deque` (Double Ended QUEue) whose
operations on either side are O(1).