[IMP] hr_attendance : add extra_hours concepts
Two extra_hours fields have been added. One for the employee, and
one for the attendances.

Employee:
---------

The amount on the employee represents the total of extra hours (+ or -)
an employee has done from his first attendance (ever in timeline) check_in to now. It is computed by
comparing effective worked hours (attendances) to the employee's
schedule (resource calendar). Leaves from the calendar are also used in
the computation.

This amount is visible with a stat button on the employee's form view and
on the "My Profile" view, and the kiosk + check in/out page.

Attendance:
-----------

The amount on an attendance represents the total of extra hours (+ or -)
for the WHOLE DAY. So every attendances of a same day will have the same amount.
It is computed with other attendances of the same day, compared to the calendar.

This amount is accessible in reports only. It isn't displayed anywhere by attendance
since every attendance of a same day have the same amount. It could be confusing otherwise.

For hr_attendance, reporting views are now SQL views.

TASK#1904850