[REF] stock: header in picking form
Only keep buttons in the header to be able to disable them with an xpath
without breaking every attrs elsewhere in the form.

Why? We want to disable these header buttons when opening a picking from
the bach transfer (by clicking on a line on a o2m list).

task-2069646