[REF] website, web_editor: move theme customization to edit mode
Previously, theme customization was done through a modal dialog in the
website module, this was not very user friendly since the rest of the
customization for the website was done from edit mode, meaning the user
had to exit edit mode to customize things such as theme colors or fonts.
It was also rather confusing to have these seemingly related things in
completely unrelated places.

This commit moves all of the options that used to be in the theme
customization modal into a new tab in the editor's left panel.

task-2088298