[IMP] stock: Zero Quantity Count
Whenever a location is theoriticaly completely empty, the user can now
either confirm the empty location or adjust the inventory.
This allows to keep a more consistent inventory in the day-to-day work by making small
inventory adjustment from times to times instead of doing big inventory
adjustments more episodically.

This commit implement this feature for the desktop `client_action` by
adding a setting in the inventory settings, and adding a
Zero Quantity Count wizard that can be eventually called during
the picking validation process.

When the pickings validated are emptying locations, this wizard is raised
and asks for a confirmation of the location's emptiness, but also offer
the possibilty to make an inventory adjustment for those locations.
This is why pickings have now a field `inventory_ids`, those inventory
will be validated right after the picking is validated.

As the `theoritical_qty` and the `difference_qty` column of the
inventory lines are wrong in such a case, those are hidden if the
inventory adjustment is made from a Zero Quantity Count.

TaskID: 1938108