[INP] web: force a master password on insecure dbs
When creating a new database, generate a random master password for
it.

The motivation for this change is to have, by default, a secure master
password set for any Odoo deployment.
Often, users do not realise that, when making an Odoo installation
accessible on internet for own usage, anyone else can also access it.