[FIX] base: avoid whole cache invalidation when removing translations.
The case is exactly the one tested by `test_sync_update` (which also
happen during database upgrade):

We update a view by writing on the `arch` (computed) field.
The field is protected while calling the inverse function.
This inverse function writes back the value on the `arch_db` field.
`arch_db` being translated, the translations are synchronized.
During the unlink of outdated translations, the whole cache is cleared
because some `on delete cascase` may have happen.
The constraint on `arch_db` are then verified.
The constraint read the field `arch`.
`arch` is not in cache anymore and should therefore be recomputed.
However, being protected this field is simply ignored and `False` is set
in the cache.

A workaround was used to not validate the view is this case but
(in addition to being wrong) is not enough in case of an inherited views
is created/updated later on.