[IMP] event: improve some views and clean the code
Purpose
=======
Improve some view and reset the templates on event mail and
clean the code.

Specifications
==============
Improve the event kanban view to be able to insert elements
in the kanban card footer.

Add a context key "default_event_id" on the ``event.mail``,
so the event is automatically set.

When writing on the notification type (mail/sms) of an event mail,
reset the template.

Make the table larger in the event category form views.

Create a method ``_get_event_mail_values`` which returns values
for ``event.mail`` automatic creation. So we can inherits in other
models.

Task #2127615