[IMP] project: introduce the possibility to create a project template
The purpose of this is to allow the creation of template projects that can be used as starting point for normal projects but don't appear
in reporting, on the dashboard, on the portal...

Tasks can also be created in those projects but no timesheets, quotations, sale orders or invoices can
be created from those tasks. We can link a partner to template tasks. They don't appear on the partner form.

On the product form, for services, the project template field is now linked to a template project and not to a normal project.

Task-1975084