[FIX] hr_attendance : remove auto kiosk mode for every user
Because menus in app attendance weren't in an appropriate order,
opening the app opened the kiosk mode directly and it wasn't possible to
do something else since other menus are hidden in kiosk mode.

Menus order has been changed, now it is the first menu for people who only have
access to the kiosk mode so that it opens correctly at once.
But for other people, it is the Check In / Check Out that shows first, so that
it prevent from opening the kiosk mode when going into the attendance app.