[IMP] project : render templates visible only to project manager.
This ir rule is created so that templates are not visible to project users.
This way, users can't create tasks in templates (from the all tasks menu).

On a product form, this rule isn't applied for the project template field
to allow users to create services with project templates even if they don't
have the rights to access the templates.

Task-1975084