[IMP] core: deduplicate BINARY_ and INPLACE_ mnemonics
They're exact mirrors of one another (with a different prefix), but we
only supported a fraction of the INPLACE bytecodes for unknown reason.

Initially looked at simply aligning the BINARY and INPLACE mnemonics
e.g.

     'BINARY_POWER',  'BINARY_MULTIPLY',  'BINARY_FLOOR_DIVIDE',
    'INPLACE_POWER', 'INPLACE_MULTIPLY', 'INPLACE_FLOOR_DIVIDE',
    ...

so it would be easier to notice discrepancies, but it seems simpler
and clearer to just move the suffixes to a single list and generate
the BINARY_ and INPLACE_ mnemonics from that