[FIX] point_of_sale: anglo saxon and tax lines in different currency
The problem occurs when the pos session has currency that is different
from the company's currency. Additionally, product that is tracked in
accounting is also sold in this session.

This situation results to anglo saxon move lines that are incorrectly
converted.

The current implementation considers the value from the stock valuation
layer to be in the session's currency. However, that is not the real
case because values in the valuation layer are always in the company
currency.

With regards to tax line, the tax_base_amount is not converted to the
company's currency. It's value should always be in company's currency.

This commit fixes the behavior of pos in generating anglo saxon lines
and setting the correct tax_base_amount in the pos session's accounting
entry.