[IMP] mail: improve my activities
People are actually using the next activities as a To Do list, and it's hard for
them to get a clear view of all their Today's activities (as it's separated per
app).

So, removed the activity icon on the systray dropdown and activated the
Calendar-Activity-Kanban-List view on the activities.

task-2153113