[IMP] survey:  Illustrate your questions and answers with media
Now you can add a description to a question in order to illustrate or describe it. This option is available
with the html field description which is now render in the backend.

You have also the possibility to add pictures on an answer for mcq if it makes sens with the
display mode, this option is available in the tab "Options" under the section "Display Mode".

Demo data has been added to test these new features.

task-2083823