[FIX] web,res_partner: add protocol (http://) if not found in url.
Currently in res.parther when we add the website, it will call the function
_clean_website() and check the URL has a protocol(http) or not. if not found
then it will add the "http" as prefix and then store it in the database.
Drawback of this solution is if we want to use the url widget field in any other
model then we have to do the redundant code.

With this solution, by adding one option website_path=True along with url
widget will redirects to the own website path and website_path=False will
consider this as the external website url and prefixed http to the URL and
redirect externally.

task-2153184