[IMP] website_slides: improve the ui on a course (documentation)
Before, uncategorized slides  weren't display correctly on the slide_channel. It was impossible
to view all the slides if there were more than 4 slides uncategorized because the "View all" link wasn't
display.

Now, when there is no section/category on a course, it displays the last 4 slides with the "View all" link
which allows to be redirected to the uncategorized category (category_id=-1). This request displays all the
uncategorized slides with a pager if the limit of 12 slides per page is exceeded.
If there is at least one category created, all uncategorized slides are displayed under a section "Uncategorized".

Additionaly, there is a new button which allows to create a category directly on the course page to allow the user to
organize is course more effectively.

task-2161391