[FIX] website_slides: Fix slide reordering
This commit fixes the reordering of slides on upload or drag and drop
in frontend.

This issue was due to the fact that in a previous commit:
  c4eba6f

uncategorized slides were moved to the bottom of the list,
which broke the reordering after drag and drop, as we use the dom order itself
to reorder the slides with a simple rpc call to dataset/resequence.

Furthermore, having the uncategorized slides at the top of the list,
allows us to be coherent with the backend behaviour.

Another issue was that, on slide upload, in the controller, we used
the category_id on the slide (which seems to not always be correct at
this stage) instead of the one given by the javascript. This caused the slide
to not be uploaded in the right section.

LINKS:

TaskID: 2090927
PR: #39086