[IMP] website: add consent duration on s_popup snippet
This commit introduce a new snippet option for s_popup to control the consent
duration in days. This value will be used to know if the popup should be shown
or not.
If the user close the popup, it won't be shown until the consent duration is
over.

This will also avoid popup to be shown on exit (display mode) more than once.
Indeed, every mouseout of the screen (changing tab, changing app, alt tab etc)
would show the popup, as well as page navigation.

task-2087003