[IMP] calendar: private events shouldn't be editable/deletable
Currently, the user will get an access rights error as soon as they try to
edit or delete the private event

Thus, hide the edit and delete button in that case. Also, hide the popover
fields. Make the 'privacy' and 'show_as' fields required to restrict the empty
option.

task-2151150