[REF] website_blog: move snippets to match website snippet organization
Recently, we changed the way snippet are organized in website to allow
for versioning of the snippets' css and ftonend js, to prevent new
changes from breaking old versions of a snippet. This commit moves all
the parts of the s_latest_posts and its big picture counterpart (which
are really just two variations of a single snippet) to match this new
organization, such that we can support css and js versioning in the
future.

task-2152188
PR: odoo/odoo#42466