[IMP] event_barcode: add misc for event_barcode
PURPOSE

Before this task you had to install event_sale to see the popup window
in the barcode interface when scanning an attendee.
The purpose of this task is to make it pop event when event_sale is not
installed.

SPECIFICATIONS

Add the minimal information about the registration in the event module
and overrides it in the event_sale module to add the necessary informations.
That makes the window popup appear everytime when scanning an attendee.

LINKS

task ID : 2093336
PR : #43420