[IMP] event_crm: create leads from attendees
This new module allows to create rules in order to create
leads from event's registrations.

A new menu, "Lead Generation", in the configuration tab of an event
allows to create a set of conditions to create a lead with some
pre-filled values as the type of lead, tags or salesperson.

There are two types of lead creation:
	-One per Attendee: create a lead for each attendee
	-One per Order: create a lead for a set of attendees
The last one is only available if it's possible to register a group
of attendees in one action (depends of website_event).

The flow of lead creation is as follow:
When attendees are created we will check if any rules applies to the attendees.
For the rule to be apply, it has to met some conditions:
	- The rule is active.
	- If the company is set on the rule, it must match.
	- If the event category and the event aren't specified, it's always OK.
	- If only the event category is set, it must match.
	- If only the event is set, it must match.
	- If the event category and the event are set, one of them must match.

When the conditions are matches, it will create the lead with some information
pre-filled (name, type, contact_name, ...)

Once a lead is created for an event registration, a stat button on the event
will be available to show the number of leads created for this event and to
display all of them in order to access them.

And on the lead form, a stat button will be display to show the attendees
linked to this lead.

task-2166679