[IMP] web: display statusbar widget always
widgets are not displayed if there is no value and isSet method of widget
returns false, it is intended behavior as we do not want to display widget
if there is no meaningful value set.

due to this behavior statusbar widget creates awkward UX change when user
moves from readonly to edit, as statusbar is added in header due to which
changing mode adds new element i.e. header if statusbar does not have meaningful
value.

To avoid this, override isSet method in FieldStatus widget and always return true,
to display statusbar widget always even when no value set and mode is readonly.

task-2172272