[FIX] stock: package location dest with putaway
When a package contains products with putaway to different locations,
it could happen the move lines locations are took from putaway location.
In this case, location dest of package and its move lines aren't
consistent anymore. With this commit, when moves are assigned, package
and its move lines `location_dest_id` are checked, and are all set to
the first move line `location_dest_id` if there is a difference.

task-2146583
task-2148494