[IMP] website_profile: use badge level image as default badge image
Before this commit, if no image was set on the badge, no image was shown when
dispalying the badge.

After this commit, if a badge level is set on the badge and no image is set,
the badge level image will be used for the badge.

Also add padding to avoid the title to touch the badge image.

Task id: 1981399
PR #43577