[IMP] mail: fix 'show older messages' link glitch
Before this commit:
after creating a ticket from the 'convert to ticket' button, it shows 'show
older messages' link in the chatter.
But nothing happens when clicking on it. It loads the messages only after
reloading the page (even if 'show older..' link clicked or not)

After this commit:
It will not show the 'show older messages' link.
It will load all the messages after the ticket creation.

task-2168287