[IMP] website_event_track: improve talk proposal inputs
PURPOSE

When creating a talk proposal, the user couldn't style his presentation
and his biography with the simple textarea input.
After this commit he'll be able to style both of them with an improved
textarea input.

SPECIFICATION

Add a special class on textarea which instantiate a wysiwyg widget to
improve the textarea input.

Task ID : 2093336
PR : #43420