[IMP] survey: add a 'live session' mode to the survey app
PURPOSE

Add a "live session" mode for survey that allows the host to interact with its audience.
He controls the pace of the survey and the audience answers questions one at a time.

Results for each question can be displayed by the host to adapt his speech in real time.

A gamification component even allows to give points to the attendees, based on the quickness of
their answer, and create ranking to keep everyone's attention.

SPECIFICATIONS

This whole feature is deeply integrated with survey.
It adds a layer of "survey session" on top if it that gathers attendees and answers during the
lifespan of the session.

The attendee identity is filled in by a special question that is marked as "Save as user nickname".
If the host has not configured that question or if the user doesn't answer, they are marked as
"Anonymous" in the rankings.

The answer score computation is also altered during the session to take the speed of the answer
into account.

As we do not have a way to simplify links and it still might be hard for attendees to fully type
"https://thenameofthecompany.com/survey/superQuiz", the best way to share the survey link will
still be to use the "Share" button on the survey form view, just like regular surveys.

FLOW

HOST point of view:

- From the Survey Form view, the host can create a "live session"
  Only one session per survey can be created at a time.
- The host can select a few options for this new session:
    - "Competitive Mode" that will allow ranking the attendees based on their score.
    - "Time Limit" that defines if questions are limited in time (+ config in seconds)
    - "Reward quick answers" that gives more points to attendees if they answer quickly
- The host clicks on "Open Session Manager" to mark the session as 'ready', meaning all attendees
  landing on the survey page will be linked to the session.
- When clicking an "Open Session Manager", he lands on a page with a recap of all the different
  session settings as well as the link for the attendee to join that session and the number of
  attendees.
- The host starts the session, activating the first question of this session.
  We keep an active reference to the "current_question_id" and use the bus to trigger an event
  that will refresh the survey page for all attendees, showing them the question and allowing them
  to answer.
- The host gets a "question management screen", from which he can:
    - See the current question text & suggested answers
    - See the current question timer
    - See the number of answers received for the question
    - Display the answers of the question (same view as the survey "results" page, but only for
      the question)
    - Display the ranking of attendees (if "competitive mode" is enabled)
- The host controls the pace of the survey by moving to the next question until it's the last one
  of the survey, then ends the current session.
  When the session is closed, we mark the answers of the attendees as "done".
- The host lands on a screen where he sees the survey results for this session only and the ranking
  of all attendees.

ATTENDEE point of view:

- He reaches the survey when a session and open, and gets a screen asking him to wait until the
  host decides to start the session.
- When the sessions starts, he's automatically redirected to the first question (see above).
- If the session is already started when he reaches the link, he lands on the current question
  (and NOT on the first one)
- If the host has configured a time limit, the attendee sees the countdown while he's answering.
- When the attendee answers the question, the screen tells him the answer is registered and he has
  to wait for the host to go to the next question, which will happen automatically.
- If the timer reaches 0 and he has not submitted his answer, it's too late and he waits for the
  next one.
- It continues like that until the end of the session.
- At the end of the session, the attendee gets a screen with his global results (same as a regular
  scored survey).

Task#1972640